WIP: monitor

+++ internal/pkg/kube/scale.go
@@ -6,12 +6,12 @@ import (
 )
 
 // Scale 伸缩Pods数量
-func (client *Client) Scale(ns, deployment *string, num int32) error {
+func (client *Client) Scale(ns, deployment *string, num uint) error {
 	deploy, err := client.clientSet.AppsV1().Deployments(*ns).Get(context.TODO(), *deployment, metav1.GetOptions{})
 	if err != nil {
 		return err
 	}
-	*deploy.Spec.Replicas = num
+	*deploy.Spec.Replicas = int32(num)
 	if _, err := client.clientSet.AppsV1().Deployments(*ns).Update(context.TODO(), deploy, metav1.UpdateOptions{}); err != nil {
 		return err
 	}
